fixed bugs

+++ router/sr_router.c
@@ -29,7 +29,7 @@ int sanity_check(uint8_t *buf, unsigned int length);
 int handle_chksum(sr_ip_hdr_t *ip_hdr);
 void construct_eth_header(uint8_t *buf, uint8_t *dst, uint8_t *src, uint16_t type);
 void construct_arp_header(uint8_t *buf, struct sr_if* source_if, sr_arp_hdr_t *arp_hdr, unsigned short type);
-void construct_ip_header(uint8_t *buf, uint8_t dst, uint8_t src, uint16_t type);
+void construct_ip_header(uint8_t *buf, uint32_t dst, uint32_t src, uint16_t type);
 uint8_t* construct_icmp_header(uint8_t *ip_buf, struct sr_if* source_if, uint8_t type, uint8_t code, unsigned long total_len);
 /*---------------------------------------------------------------------
  * Method: sr_init(void)
@@ -265,7 +265,7 @@ void construct_arp_header(uint8_t *buf, struct sr_if* source_if, sr_arp_hdr_t *a
   reply_arp_hdr->ar_tip = arp_hdr->ar_sip;
 }
 
-void construct_ip_header(uint8_t *buf, uint8_t dst, uint8_t src, uint16_t type) {
+void construct_ip_header(uint8_t *buf, uint32_t dst, uint32_t src, uint16_t type) {
   sr_ip_hdr_t *ip_hdr = (sr_ip_hdr_t *)(buf);
   ip_hdr->ip_src = src;
   ip_hdr->ip_dst = dst;
